refac: Move NullAuth into client module
